Safari-609.1.11.1/LayoutTests/fast/media/mq-any-hover-matchMedia-expected.txt

Test the any-hover media feature with machMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.matchMedia("(any-hover)").matches = true
window.matchMedia("(any-hover: hover)").matches = true
window.matchMedia("(any-hover: on-demand)").matches = false
window.matchMedia("(any-hover: none)").matches = false
window.matchMedia("(any-hover: Hover)").matches = true
window.matchMedia("(any-hover: On-demand)").matches = false
window.matchMedia("(any-hover: None)").matches = false
window.matchMedia("(Any-Hover: hover)").matches = true
window.matchMedia("(Any-Hover: on-demand)").matches = false
window.matchMedia("(Any-Hover: none)").matches = false
PASS successfullyParsed is true

TEST COMPLETE
